# Server/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/doctors": {
      "get": {
        "tags": ["doctors"],
        "summary": "Get Doctors",
        "operationId": "get_doctors_doctors_get",
        "parameters": [
          {
            "name": "max",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 5,
              "title": "Max"
            }
          },
          {
            "name": "currPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Currpage"
            }
          },
          {
            "name": "searchQuery",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Searchquery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DoctorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "": ""
    },
    "/doctors/{id}": {
      "get": {
        "tags": ["doctors"],
        "summary": "Get Doctor",
        "operationId": "get_doctor_doctors__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Doctor" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "": ""
    },
...
  },
  "components": {
    "schemas": {
...
      "Clinic": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "contact": { "type": "string", "title": "Contact" },
          "address": { "type": "string", "title": "Address" },
          "facilities": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Facilities",
            "default": []
          },
          "location": {
            "anyOf": [
              { "$ref": "#/components/schemas/Location" },
              { "type": "null" }
            ]
          },
          "whatsapp": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Whatsapp"
          },
          "parking_available": {
            "type": "boolean",
            "title": "Parking Available",
            "default": false
          }
        },
        "type": "object",
        "required": ["name", "contact", "address"],
        "title": "Clinic"
      },
      "Doctor": {
        "properties": {
          "status": { "type": "string", "title": "Status" },
          "fee": { "$ref": "#/components/schemas/Fee" },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "default": false
          },
          "reviews": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Reviews"
          },
          "rating": {
            "anyOf": [
              { "type": "integer" },
              { "type": "number" },
              { "type": "null" }
            ],
            "title": "Rating"
          },
          "consults_online": {
            "type": "boolean",
            "title": "Consults Online",
            "default": false
          },
          "booking_enabled": {
            "type": "boolean",
            "title": "Booking Enabled",
            "default": false
          },
          "base_fee": { "type": "integer", "title": "Base Fee" },
          "base_consult_time": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Base Consult Time"
          },
          "experience": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Experience"
          },
          "secondary_specializations": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Secondary Specializations",
            "default": []
          },
          "currently_available": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Currently Available"
          },
          "last_updated": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Last Updated"
          },
          "next_available": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Next Available"
          },
          "office": {
            "anyOf": [
              { "$ref": "#/components/schemas/Clinic" },
              { "type": "null" }
            ]
          },
          "schedules": {
            "items": { "$ref": "#/components/schemas/Schedule" },
            "type": "array",
            "title": "Schedules",
            "default": []
          },
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "email": { "type": "string", "format": "email", "title": "Email" },
          "credentials": { "type": "string", "title": "Credentials" },
          "primary_specialization": {
            "type": "string",
            "title": "Primary Specialization"
          }
        },
        "type": "object",
        "required": [
          "status",
          "fee",
          "base_fee",
          "name",
          "email",
          "credentials",
          "primary_specialization"
        ],
        "title": "Doctor"
      },
      "DoctorResponse": {
        "properties": {
          "doctors": {
            "items": { "$ref": "#/components/schemas/Doctor" },
            "type": "array",
            "title": "Doctors"
          },
          "curr_count": { "type": "integer", "title": "Curr Count" },
          "total_count": { "type": "integer", "title": "Total Count" }
        },
        "type": "object",
        "required": ["doctors", "curr_count", "total_count"],
        "title": "DoctorResponse"
      },
      "Fee": {
        "properties": {
          "online": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Online"
          },
          "in_person": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "In Person"
          }
        },
        "type": "object",
        "required": ["online", "in_person"],
        "title": "Fee"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Location": {
        "properties": {
          "lat": { "type": "number", "title": "Lat" },
          "lng": { "type": "number", "title": "Lng" }
        },
        "type": "object",
        "required": ["lat", "lng"],
        "title": "Location"
      },
      "Schedule": {
        "properties": {
          "weekday": { "type": "string", "title": "Weekday" },
          "slots": {
            "items": { "$ref": "#/components/schemas/Slot" },
            "type": "array",
            "title": "Slots"
          },
          "clinic": {
            "anyOf": [
              { "$ref": "#/components/schemas/Clinic" },
              { "type": "null" }
            ]
          },
          "end": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "End"
          },
          "start": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Start"
          },
          "hours_available": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Hours Available"
          }
        },
        "type": "object",
        "required": ["weekday", "slots"],
        "title": "Schedule"
      },
      "Slot": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "begin": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Begin"
          },
          "duration": { "type": "integer", "title": "Duration" },
          "booked": { "type": "boolean", "title": "Booked", "default": false },
          "mode": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": ["duration"],
        "title": "Slot"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}